requirements.txt
whisper
numpy
moviepy
Pillow
